move bip ce model to its own package
