+++ controller/src/main/java/gov/va/vro/controller/BipController.java
@@ -9,9 +9,9 @@ import gov.va.vro.api.responses.BipContentionUpdateResponse;
 import gov.va.vro.api.responses.BipFileUploadResponse;
 import gov.va.vro.model.bip.BipClaim;
 import gov.va.vro.model.bip.BipCreateClaimContentionPayload;
-import gov.va.vro.model.bip.BipFileProviderData;
-import gov.va.vro.model.bip.BipFileUploadPayload;
-import gov.va.vro.model.bip.BipFileUploadResp;
+import gov.va.vro.model.bipevidence.BipFileProviderData;
+import gov.va.vro.model.bipevidence.BipFileUploadPayload;
+import gov.va.vro.model.bipevidence.BipFileUploadResp;
 import gov.va.vro.model.bip.BipUpdateClaimContentionPayload;
 import gov.va.vro.model.bip.BipUpdateClaimPayload;
 import gov.va.vro.model.bip.BipUpdateClaimResp;
