+++ service/provider/src/main/java/gov/va/vro/service/provider/bip/service/BipClaimService.java
@@ -1,7 +1,7 @@
 package gov.va.vro.service.provider.bip.service;
 
-import gov.va.vro.model.bip.BipFileProviderData;
-import gov.va.vro.model.bip.BipFileUploadPayload;
+import gov.va.vro.model.bipevidence.BipFileProviderData;
+import gov.va.vro.model.bipevidence.BipFileUploadPayload;
 import gov.va.vro.model.bip.ClaimContention;
 import gov.va.vro.model.bip.ClaimStatus;
 import gov.va.vro.model.bip.FileIdType;
