+++ shared/domain/src/main/java/gov/va/vro/model/bipevidence/BipFileUploadResp.java
@@ -1,4 +1,4 @@
-package gov.va.vro.model.bip;
+package gov.va.vro.model.bipevidence;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
